stress-opcode: cast integers to double before double division

+++ stress-opcode.c
@@ -631,7 +631,7 @@ finish:
 	    stress_instance_zero(args) &&
 	    (args->instances > 0)) {
 		const double secs_in_tropical_year = 365.2422 * 24.0 * 60.0 * 60.0;
-		double estimated_duration = (duration * num_opcodes / vstate->ops_attempted) / args->instances;
+		double estimated_duration = (duration * (double)num_opcodes / (double)vstate->ops_attempted) / (double)args->instances;
 
 		if (estimated_duration > secs_in_tropical_year * 5) {
 			estimated_duration = round(estimated_duration / secs_in_tropical_year);
